+++ bindings/python/pyonmttok/__init__.py
@@ -1,0 +1,1 @@
+from pyonmttok.tokenizer import Tokenizer